9acc0606390355c1a95ac23a2c5c1114f7a3b447 player: add building blocks for playback quirks
based off the OpenMPT code.

Saga Musix suggests to clear the quirks when the song is saved, so that
old incompatible modules don't "grandfather in" to newer versions. I think
that's definitely better than saving them (and possibly growing the resulting
file size), so that's what I'm going to implement

OpenMPT itself saves the state of the quirks. It is possible that we could
load and parse this data, so that they will play the same within both
players.
---------
Added generic type BitArray<T> for any T an int-based Enum.

Co-Authored-By: Paper <paper@tflc.us>