16764cc13541c240a34c387710f28335c66868c4 player/mixer: backport wraparound destination checking from openmpt
it still doesn't work right though
(reported by cs127: coming out of a sustain loop causes a buffer
overflow SOMETIMES. other times it doesn't, and the logic that causes
it is definitely in the loop wrap-around code, because it doesn't
happen if i comment it out. i still don't really know what's going on
there so i'm just gonna put it off for now)

actually, what's the deal with the CHN_SUSTAINLOOP flag? it's being
checked, but does it actually get unset if the sustain loop is
finished playing? what happens if we have a normal loop too?
Co-Authored-By: Paper <paper@tflc.us>